requirements.txt
requests
psycopg2-binary
numpy
pandas